@@ -2,36 +2,32 @@
 
 namespace Asynit;
 
-use Amp\Artax\DefaultClient;
-use function Amp\call;
-use Amp\Promise;
-use Amp\Socket\ClientTlsContext;
-use Amp\Sync\Lock;
-use Amp\Sync\Semaphore;
 use Asynit\Assert\AssertWebCaseTrait;
-use Asynit\HttpClient\ArtaxAsyncAdapter;
-use Http\Client\HttpAsyncClient;
-use Http\Message\MessageFactory;
+use Psr\Http\Client\ClientInterface;
+use Psr\Http\Message\RequestFactoryInterface;
+use Amp\Sync\Semaphore;
 use Psr\Http\Message\RequestInterface;
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\StreamFactoryInterface;
 
 class TestCase
 {
     use AssertWebCaseTrait;
 
-    /** @var MessageFactory */
-    private $messageFactory;
-
-    /** @var Semaphore */
-    private $semaphore;
-
-    /** @var HttpAsyncClient */
-    private $client;
-
-    final public function __construct(MessageFactory $messageFactory, Semaphore $semaphore, Test $test)
+    final public function __construct(
+        private RequestFactoryInterface $requestFactory,
+        private StreamFactoryInterface $streamFactory,
+        private Semaphore $semaphore,
+        Test $test,
+        protected ClientInterface $client,
+    )
     {
-        $this->messageFactory = $messageFactory;
-        $this->semaphore = $semaphore;
         $this->test = $test;
+    }
+
+    public function initialize()
+    {
+        $this->client = $this->setUp($this->client);
     }
 
     /**
@@ -39,108 +35,77 @@ class TestCase
      *
      * Allow to set default services and context, and also decorate the http async client.
      *
-     * @return \Generator|Promise|HttpAsyncClient
+     * @return ClientInterface
      */
-    public function setUp(HttpAsyncClient $asyncClient)
+    public function setUp(ClientInterface $asyncClient): ClientInterface
     {
         return $asyncClient;
-    }
-
-    final public function initialize(bool $allowSelfSignedCertificate = false)
-    {
-        return call(function () use ($allowSelfSignedCertificate) {
-            $this->client = yield call(function () use ($allowSelfSignedCertificate) {
-                if ($allowSelfSignedCertificate) {
-                    $tlsContext = new ClientTlsContext();
-                    $tlsContext = $tlsContext->withoutPeerVerification();
-                }
-
-                return $this->setUp(new ArtaxAsyncAdapter($this->messageFactory, new DefaultClient(null, null, $tlsContext ?? null)));
-            });
-        });
     }
 
     /**
      * Allow to test a rejection or a resolution of an async call.
      */
-    final protected function sendRequest(RequestInterface $request): Promise
+    final protected function sendRequest(RequestInterface $request): ResponseInterface
     {
-        return \Amp\call(function () use ($request) {
-            /** @var Lock $lock */
-            $lock = yield $this->semaphore->acquire();
-            $response = yield $this->client->sendAsyncRequest($request);
+        $lock = $this->semaphore->acquire();
+        $response = $this->client->sendRequest($request);
+        $lock->release();
 
-            $lock->release();
-
-            return $response;
-        });
+        return $response;
     }
 
-    /**
-     * @param        $uri
-     * @param array  $headers
-     * @param null   $body
-     * @param string $version
-     */
-    final protected function get($uri, $headers = [], $body = null, $version = '1.1'): Promise
+    final protected function get(string $uri, array $headers = [], $body = null, ?string $version = null): ResponseInterface
     {
-        return $this->sendRequest($this->messageFactory->createRequest('GET', $uri, $headers, $body, $version));
+        return $this->sendRequest($this->createRequest('GET', $uri, $headers, $body, $version));
+    }
+    
+    final protected function post(string $uri, array $headers = [], $body = null, ?string $version = null): ResponseInterface
+    {
+        return $this->sendRequest($this->createRequest('POST', $uri, $headers, $body, $version));
     }
 
-    /**
-     * @param        $uri
-     * @param array  $headers
-     * @param null   $body
-     * @param string $version
-     */
-    final protected function post($uri, $headers = [], $body = null, $version = '1.1'): Promise
+    final protected function patch(string $uri, array $headers = [], $body = null, ?string $version = null): ResponseInterface
     {
-        return $this->sendRequest($this->messageFactory->createRequest('POST', $uri, $headers, $body, $version));
+        return $this->sendRequest($this->createRequest('PATCH', $uri, $headers, $body, $version));
     }
 
-    /**
-     * @param        $uri
-     * @param array  $headers
-     * @param null   $body
-     * @param string $version
-     */
-    final protected function patch($uri, $headers = [], $body = null, $version = '1.1'): Promise
+    final protected function put(string $uri, array $headers = [], $body = null, ?string $version = null): ResponseInterface
     {
-        return $this->sendRequest($this->messageFactory->createRequest('PATCH', $uri, $headers, $body, $version));
+        return $this->sendRequest($this->createRequest('PUT', $uri, $headers, $body, $version));
     }
 
-    /**
-     * @param        $uri
-     * @param array  $headers
-     * @param null   $body
-     * @param string $version
-     *
-     * @return Promise
-     */
-    final protected function put($uri, $headers = [], $body = null, $version = '1.1')
+    final protected function delete(string $uri, array $headers = [], $body = null, ?string $version = null): ResponseInterface
     {
-        return $this->sendRequest($this->messageFactory->createRequest('PUT', $uri, $headers, $body, $version));
+        return $this->sendRequest($this->createRequest('DELETE', $uri, $headers, $body, $version));
     }
 
-    /**
-     * @param        $uri
-     * @param array  $headers
-     * @param null   $body
-     * @param string $version
-     */
-    final protected function delete($uri, $headers = [], $body = null, $version = '1.1'): Promise
+    final protected function options(string $uri, array $headers = [], $body = null, ?string $version = null): ResponseInterface
     {
-        return $this->sendRequest($this->messageFactory->createRequest('DELETE', $uri, $headers, $body, $version));
+        return $this->sendRequest($this->createRequest('OPTIONS', $uri, $headers, $body, $version));
     }
 
-    /**
-     * @param        $uri
-     * @param array  $headers
-     * @param null   $body
-     * @param string $version
-     */
-    final protected function options($uri, $headers = [], $body = null, $version = '1.1'): Promise
+    protected function createUri(string $uri): string
     {
-        return $this->sendRequest($this->messageFactory->createRequest('OPTIONS', $uri, $headers, $body, $version));
+        return $uri;
+    }
+
+    private function createRequest(string $method, string $uri, array $headers = [], $body = null, ?string $version = null): RequestInterface {
+        $request = $this->requestFactory->createRequest($method, $this->createUri($uri));
+
+        foreach ($headers as $name => $value) {
+            $request = $request->withHeader($name, $value);
+        }
+
+        if ($body !== null) {
+            $body = $this->streamFactory->createStream($body);
+
+            $request = $request->withBody($body);
+        }
+
+        if ($version !== null) {
+            $request = $request->withProtocolVersion($version);
+        }
+
+        return $request;
     }
 }
