@@ -2,14 +2,17 @@
 
 namespace Asynit\Command;
 
-use Amp\Loop;
+
 use Asynit\Output\OutputFactory;
 use Asynit\Parser\TestPoolBuilder;
 use Asynit\Parser\TestsFinder;
 use Asynit\Runner\PoolRunner;
 use Asynit\TestWorkflow;
 use Doctrine\Common\Annotations\AnnotationReader;
+use GuzzleHttp\Psr7\HttpFactory;
 use Http\Message\MessageFactory\GuzzleMessageFactory;
+use Nyholm\Psr7\Factory\Psr17Factory;
+use Revolt\EventLoop;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
@@ -57,14 +60,12 @@ class AsynitCommand extends Command
 
         // Build services for parsing and running tests
         $builder = new TestPoolBuilder(new AnnotationReader());
-        $runner = new PoolRunner(new GuzzleMessageFactory(), new TestWorkflow($chainOutput), $input->getOption('concurrency'));
+        $factory = new HttpFactory();
+        $runner = new PoolRunner($factory, $factory, $factory, new TestWorkflow($chainOutput), $input->getOption('concurrency'));
 
         // Build a list of tests from the directory
         $pool = $builder->build($testMethods);
-
-        Loop::run(function () use ($runner, $pool) {
-            $runner->loop($pool);
-        });
+        $runner->loop($pool);
 
         // Return the number of failed tests
         return $countOutput->getFailed();
