@@ -4,7 +4,6 @@ namespace Asynit\Tests;
 
 use Asynit\Annotation\Depend;
 use Asynit\TestCase;
-use Http\Client\Exception;
 use Psr\Http\Message\ResponseInterface;
 
 class FunctionalTests extends TestCase
@@ -16,7 +15,7 @@ class FunctionalTests extends TestCase
 
     public function testGet()
     {
-        $response = yield $this->get('http://127.0.0.1:8081');
+        $response = $this->get('https://httpbin.org/');
 
         $this->assertInstanceOf(ResponseInterface::class, $response);
         $this->assertStatusCode(200, $response);
@@ -29,24 +28,21 @@ class FunctionalTests extends TestCase
         $exception = null;
 
         try {
-            $response = yield $this->get('http://something-is-not-reachable');
+            $response = $this->get('http://something-is-not-reachable');
         } catch (\Exception $e) {
             $exception = $e;
         }
 
         $this->assertNotNull($exception, 'Not null exception');
-        $this->assertInstanceOf(Exception::class, $exception);
     }
 
-    /**
-     * @Depend("testGet")
-     */
+    #[Depend("testGet")]
     public function testDepend($value)
     {
         $this->assertSame('foo', $value);
     }
 
-    /** @Depend("Asynit\Tests\AnotherTest::test_from_another_file") */
+    #[Depend("Asynit\Tests\AnotherTest::test_from_another_file")]
     public function testDependFromAnotherFile($value)
     {
         $this->assertSame('Asynit\Tests\AnotherTest::test_from_another_file', $value);
@@ -59,31 +55,29 @@ class FunctionalTests extends TestCase
 
     public function testParallel1()
     {
-        yield $this->get('http://127.0.0.1:8081/delay/7');
+        $this->get('https://httpbin.org//delay/7');
     }
 
     public function testParallel2()
     {
-        yield $this->get('http://127.0.0.1:8081/delay/7');
+        $this->get('https://httpbin.org//delay/7');
     }
 
     public function testParallel3()
     {
-        yield $this->get('http://127.0.0.1:8081/delay/7');
+        $this->get('https://httpbin.org//delay/7');
     }
 
     public function testParallel4()
     {
-        yield $this->get('http://127.0.0.1:8081/delay/7');
+        $this->get('https://httpbin.org//delay/7');
     }
 
-    /**
-     * @Depend("testStartParallel")
-     * @Depend("testParallel1")
-     * @Depend("testParallel2")
-     * @Depend("testParallel3")
-     * @Depend("testParallel4")
-     */
+    #[Depend("testStartParallel")]
+    #[Depend("testParallel1")]
+    #[Depend("testParallel2")]
+    #[Depend("testParallel3")]
+    #[Depend("testParallel4")]
     public function testEndParallel($start)
     {
         $end = time();
@@ -96,7 +90,7 @@ class FunctionalTests extends TestCase
         return 'a';
     }
 
-    /** @Depend("get_a") */
+    #[Depend("get_a")]
     public function get_b($a)
     {
         $this->assertSame('a', $a);
@@ -104,21 +98,17 @@ class FunctionalTests extends TestCase
         return 'b';
     }
 
-    /**
-     * @Depend("get_a")
-     * @Depend("get_b")
-     */
+    #[Depend("get_a")]
+    #[Depend("get_b")]
     public function test_c($a, $b)
     {
         $this->assertSame('a', $a);
         $this->assertSame('b', $b);
     }
 
-    /**
-     * @Depend("get_a")
-     * @Depend("get_b")
-     * @Depend("Asynit\Tests\AnotherTest::get_d")
-     */
+    #[Depend("get_a")]
+    #[Depend("get_b")]
+    #[Depend("Asynit\Tests\AnotherTest::get_d")]
     public function test_c_with_d($a, $b, $d)
     {
         $this->assertSame('a', $a);
