@@ -3,27 +3,33 @@
 namespace Asynit\Tests;
 
 use Asynit\TestCase;
+use GuzzleHttp\Psr7\HttpFactory;
+use Http\Client\Common\Plugin\AddPathPlugin;
 use Http\Client\Common\Plugin\BaseUriPlugin;
 use Http\Client\Common\PluginClient;
-use Http\Client\HttpAsyncClient;
-use Http\Message\UriFactory\GuzzleUriFactory;
+use Psr\Http\Client\ClientInterface;
 use Psr\Http\Message\ResponseInterface;
 
 class HttpClientOverrideTest extends TestCase
 {
-    public function setUp(HttpAsyncClient $asyncClient): HttpAsyncClient
+    public function setUp(ClientInterface $client): ClientInterface
     {
-        $uri = (new GuzzleUriFactory())->createUri('http://127.0.0.1:8081');
+        $uri = (new HttpFactory())->createUri('https://httpbin.org/delay');
 
-        return new PluginClient($asyncClient, [
-            new BaseUriPlugin($uri),
+        return new PluginClient($client, [
+            new AddPathPlugin($uri),
         ]);
     }
 
     public function testFoo2()
     {
-        $response = yield $this->get('/delay/3');
+        $response = $this->get('/3');
 
         $this->assertInstanceOf(ResponseInterface::class, $response);
+    }
+
+    protected function createUri(string $uri): string
+    {
+        return 'https://httpbin.org' . $uri;
     }
 }
